Create a new file on this branch called 0000-01-02-arturlferreira.md
Create a new file on this branch, in a _posts folder called 0000-01-02-arturlferreira.md. You can do so using this shortcut or manually as follows:
Return to the "Code" tab
In the branch drop-down, select "myslide"
Click Create new file
In the "file name" field, type _posts/0000-01-02-arturlferreira.md. Entering the / in the filename will automatically place your file in the _posts directory.